Moved subdag_operator.py to subdag.py (#11178)
