+++ airflow/example_dags/example_subdag_operator.py
@@ -22,7 +22,7 @@
 from airflow import DAG
 from airflow.example_dags.subdags.subdag import subdag
 from airflow.operators.dummy_operator import DummyOperator
-from airflow.operators.subdag_operator import SubDagOperator
+from airflow.operators.subdag import SubDagOperator
 from airflow.utils.dates import days_ago
 
 DAG_NAME = 'example_subdag_operator'
