+++ tests/dags/test_subdag.py
@@ -25,7 +25,7 @@ from datetime import datetime, timedelta
 
 from airflow.models.dag import DAG
 from airflow.operators.dummy_operator import DummyOperator
-from airflow.operators.subdag_operator import SubDagOperator
+from airflow.operators.subdag import SubDagOperator
 
 DAG_NAME = 'test_subdag_operator'
