+++ tests/models/test_dagbag.py
@@ -320,7 +320,7 @@ class TestDagBag(unittest.TestCase):
 
             from airflow.models import DAG
             from airflow.operators.dummy_operator import DummyOperator
-            from airflow.operators.subdag_operator import SubDagOperator
+            from airflow.operators.subdag import SubDagOperator
 
             dag_name = 'master'
             default_args = {'owner': 'owner1', 'start_date': datetime.datetime(2016, 1, 1)}
@@ -371,7 +371,7 @@ class TestDagBag(unittest.TestCase):
 
             from airflow.models import DAG
             from airflow.operators.dummy_operator import DummyOperator
-            from airflow.operators.subdag_operator import SubDagOperator
+            from airflow.operators.subdag import SubDagOperator
 
             dag_name = 'master'
             default_args = {'owner': 'owner1', 'start_date': datetime.datetime(2016, 1, 1)}
@@ -491,7 +491,7 @@ class TestDagBag(unittest.TestCase):
 
             from airflow.models import DAG
             from airflow.operators.dummy_operator import DummyOperator
-            from airflow.operators.subdag_operator import SubDagOperator
+            from airflow.operators.subdag import SubDagOperator
 
             dag_name = 'nested_cycle'
             default_args = {'owner': 'owner1', 'start_date': datetime.datetime(2016, 1, 1)}
